Modifed Deployment process
 - Add README.md for instruction on properly deploying
 - frontend now uses python to run
 - added varaibles for binding backend and frontend IPs
